* Changed the type of EpisodeRecord.SiteRating from int? to decimal? * Changed the type of UserRatings.Rating from int? to decimal?

+++ src/TvDbSharper/Clients/Episodes/Json/EpisodeRecord.cs
@@ -53,7 +53,7 @@
 
         public string ShowUrl { get; set; }
 
-        public int? SiteRating { get; set; }
+        public decimal? SiteRating { get; set; }
 
         public int? SiteRatingCount { get; set; }
